+++ src/resource/solidDataset.ts
@@ -59,6 +59,7 @@ import { thingAsMarkdown, getThingAll, getThing } from "../thing/thing";
 import {
   internal_getReadableValue,
   internal_toNode,
+  internal_withChangeLog,
 } from "../thing/thing.internal";
 import { getIriAll } from "../thing/get";
 
@@ -217,18 +218,20 @@ export async function saveSolidDatasetAt<Dataset extends SolidDataset>(
     ...options,
   };
 
-  const requestInit = isUpdate(solidDataset, url)
-    ? await prepareSolidDatasetUpdate(solidDataset)
-    : await prepareSolidDatasetCreation(solidDataset);
+  const datasetWithChangelog = internal_withChangeLog(solidDataset);
+
+  const requestInit = isUpdate(datasetWithChangelog, url)
+    ? await prepareSolidDatasetUpdate(datasetWithChangelog)
+    : await prepareSolidDatasetCreation(datasetWithChangelog);
 
   const response = await config.fetch(url, requestInit);
 
   if (internal_isUnsuccessfulResponse(response)) {
-    const diagnostics = isUpdate(solidDataset, url)
+    const diagnostics = isUpdate(datasetWithChangelog, url)
       ? "The changes that were sent to the Pod are listed below.\n\n" +
-        changeLogAsMarkdown(solidDataset)
+        changeLogAsMarkdown(datasetWithChangelog)
       : "The SolidDataset that was sent to the Pod is listed below.\n\n" +
-        solidDatasetAsMarkdown(solidDataset);
+        solidDatasetAsMarkdown(datasetWithChangelog);
     throw new FetchError(
       `Storing the Resource at [${url}] failed: [${response.status}] [${response.statusText}].\n\n` +
         diagnostics,
@@ -244,7 +247,7 @@ export async function saveSolidDatasetAt<Dataset extends SolidDataset>(
   const storedDataset: Dataset &
     WithChangeLog &
     WithServerResourceInfo = Object.assign(
-    internal_cloneResource(solidDataset),
+    internal_cloneResource(datasetWithChangelog),
     {
       internal_changeLog: { additions: [], deletions: [] },
       internal_resourceInfo: resourceInfo,