+++ CHANGELOG.md
@@ -21,7 +21,7 @@ The following sections document changes that have been released already:
 ### Bugs fixed
 
 - Files sent to a Pod via `overwriteFile` or `saveFileInContainer` without a known content type
-  were rejected by Inrupt's Enterprise Solid Server were rejected with a 400 Bad Request, as
+  were rejected by Inrupt's Enterprise Solid Server with a 400 Bad Request, as
   the Solid specification says it should do. To avoid this, solid-client now sets the content type
   to `application/octet-stream` by default if no content type is known for the given file.