+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The following changes have been implemented but not released yet:
 
 ## [Unreleased]
 
+- Added documentation about the `contentType` option for file writing/saving functions.
+
 The following sections document changes that have been released already:
 
 ## [1.3.0] - 2021-01-07